chore: build no-docker

+++ backend/build.sh
@@ -1,5 +1,19 @@
 #!/bin/bash
 
+# Options
+NO_DOCKER=""
+for i in "$@"
+do
+case $i in
+  --no-docker*)
+  NO_DOCKER="true"
+  shift
+  ;;
+  *)
+  ;;
+esac
+done
+
 if [ ! -e node_modules ]
 then
   mkdir node_modules
@@ -38,45 +52,19 @@ function prop {
 }
 
 clean () {
-  docker-compose run --rm -u "$USER_UID:$GROUP_GID" gradle gradle clean
+  if [ "$NO_DOCKER" = "true" ] ; then
+    gradle clean
+  else
+    docker-compose run --rm -u "$USER_UID:$GROUP_GID" gradle gradle clean
+  fi
 }
 
-# buildNode () {
-#   #jenkins
-#   echo "[buildNode] Get branch name from jenkins env..."
-#   BRANCH_NAME=`echo $GIT_BRANCH | sed -e "s|origin/||g"`
-#   if [ "$BRANCH_NAME" = "" ]; then
-#     echo "[buildNode] Get branch name from git..."
-#     BRANCH_NAME=`git branch | sed -n -e "s/^\* \(.*\)/\1/p"`
-#   fi
-#   if [ "$BRANCH_NAME" = "" ]; then
-#     echo "[buildNode] Branch name should not be empty!"
-#     exit -1
-#   fi
-
-#   if [ "$BRANCH_NAME" = 'master' ]; then
-#       echo "[buildNode] Use entcore version from package.json ($BRANCH_NAME)"
-#       case `uname -s` in
-#         MINGW*)
-#           docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "yarn install --no-bin-links && npm update entcore && node_modules/gulp/bin/gulp.js build"
-#           ;;
-#         *)
-#           docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "yarn install && npm update entcore && node_modules/gulp/bin/gulp.js build"
-#       esac
-#   else
-#       echo "[buildNode] Use entcore tag $BRANCH_NAME"
-#       case `uname -s` in
-#         MINGW*)
-#           docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "yarn install --no-bin-links && npm rm --no-save entcore && yarn install --no-save entcore@$BRANCH_NAME && node_modules/gulp/bin/gulp.js build"
-#           ;;
-#         *)
-#           docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "yarn install && npm rm --no-save entcore && yarn install --no-save entcore@$BRANCH_NAME && node_modules/gulp/bin/gulp.js build"
-#       esac
-#   fi
-# }
-
 buildGradle () {
-  docker-compose run --rm -u "$USER_UID:$GROUP_GID" gradle gradle shadowJar install publishToMavenLocal
+  if [ "$NO_DOCKER" = "true" ] ; then
+    gradle shadowJar install publishToMavenLocal
+  else
+    docker-compose run --rm -u "$USER_UID:$GROUP_GID" gradle gradle shadowJar install publishToMavenLocal
+  fi
 }
 
 publish () {
@@ -87,22 +75,12 @@ publish () {
     echo "sonatypeUsername=$NEXUS_SONATYPE_USERNAME" >> "?/.gradle/gradle.properties"
     echo "sonatypePassword=$NEXUS_SONATYPE_PASSWORD" >> "?/.gradle/gradle.properties"
   fi
-  docker-compose run --rm -u "$USER_UID:$GROUP_GID" gradle gradle publish
+  if [ "$NO_DOCKER" = "true" ] ; then
+    gradle publish
+  else
+    docker-compose run --rm -u "$USER_UID:$GROUP_GID" gradle gradle publish
+  fi
 }
-
-# buildStatic () {
-#     docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "yarn install --no-save ode-bootstrap-neo@feat-explorer && npm run dev:build"
-# }
-
-# watch () {
-#   BUILD_APP="$(prop 'modowner')~$(prop 'modname')~$(prop 'version')"
-#   echo "Watching app $BUILD_APP"
-#   docker-compose run \
-#     --rm \
-#     -u "$USER_UID:$GROUP_GID" \
-#     -v $PWD/../$SPRINGBOARD:/home/node/$SPRINGBOARD \
-#     node sh -c "npm run watch --springboard=/home/node/$SPRINGBOARD --app=\"$BUILD_APP\""
-# }
 
 for param in "$@"
 do
@@ -110,18 +88,12 @@ do
     clean)
       clean
       ;;
-    # buildNode)
-    #   buildNode
-    #   ;;
     buildGradle)
       buildGradle
       ;;
     install)
       buildGradle
       ;;
-    # watch)
-    #   watch
-    #   ;;
     publish)
       publish
       ;;
